refactor: massive refactor for better architecture
