+++ src/app/components/filters/filters.ts
@@ -1,20 +1,20 @@
-import { create } from '../../helpers/helpFunction';
-import { isHTMLElem } from '../../helpers/helpFunction';
 import { Search } from './search';
 import { FilterWithoutSearch } from './filterWithoutSearch';
-export class Filter {
-  body: HTMLElement;
+export class Filters {
+  container: HTMLElement;
   search: Search;
-  constructor(body: HTMLElement) {
-    this.body = body;
-    this.search = new Search(body);
+  filters: FilterWithoutSearch;
+  constructor() {
+    this.container = document.createElement('aside');
+    this.container.classList.add('main-page__filter');
+    this.search = new Search();
+    this.filters = new FilterWithoutSearch();
+
+    this.container.append(this.search.container, this.filters.container);
+    this.createAll();
   }
   createAll(): void {
-    this.search.createAll();
-    const bodyForFilt = create('div', 'main-page_filter-conteiner', this.body);
-    const filterWithoutSearch = new FilterWithoutSearch(bodyForFilt);
+    const filterWithoutSearch = new FilterWithoutSearch();
     filterWithoutSearch.createAll();
   }
 }
-const filterBody = document.querySelector('.main-page__filter');
-export const filter = new Filter(isHTMLElem(filterBody));
