+++ src/index.html
@@ -8,57 +8,6 @@
     <title>Online-store</title>
   </head>
   <body>
-    <!-- <header class="header">
-      <div class="container">
-        <nav class="header__nav">
-          <ul class="header__list">
-            <li class="header__item">
-              <a href="#"><h1 class="header__logo">Music <span class="header__logo-span">World</span></h1></a
-              >
-            </li>
-            <li class="header__item">
-              <p class="header__total">Total: $<span class="header__total-amount">4514</span></p>
-              <a class="header__basket" href="#">
-                <img src="./assets/img/basket.png" alt="shopping basket" />
-                <div class="basket-score">242</div></a>
-            </li>
-          </ul>
-        </nav>
-      </div>
-    </header> -->
-    <main class="main">
-      <div class="container">
-        <section class="main-page">
-          <aside class="main-page__filter"></aside>
-          <section class="main-page__all-products">
-            <!-- <div class="main-page__header">
-              <div class="main-page__allSort">
-                <div class="main-page__configuratin">
-                  <img
-                    class="main-page__small-configuration"
-                    src="./assets/img/smallConfiguration.png"
-                    alt="small configuration"
-                  />
-                  <img
-                    class="main-page__big-configuration"
-                    src="./assets/img/bigConfiguration.png"
-                    alt="big configuration"
-                  />
-                </div>
-                <div class="main-page__sort">
-                  <p class="main-page__sort-tittle">Sort:</p>
-                  <div class="main-page__sort-selector">
-                    <p>price</p>
-                    <img class="main-page__sort-img" src="./assets/img/array.png" alt="array" />
-                  </div>
-                  <img class="main-page__sort-direction" src="./assets/img/direction.png" alt="" />
-                </div>
-              </div>
-              <p class="main-page__results"><span class="main-page__result-number">23</span> search results</p>
-            </div> -->
-          </section>
-        </section>
-      </div>
-    </main>
+
   </body>
 </html>
