+++ src/scss/_gallery.scss
@@ -13,6 +13,14 @@
     background-color: white;
     width: 290px;
     color: #494949;
+    box-shadow: 0px 0px 12px -2px rgba(156, 156, 156, 0.49);
+    border-radius: 5px;
+    transition: box-shadow 0.5s ease;
+  }
+
+  &__item:hover {
+    cursor: pointer;
+    box-shadow: 0px 4px 14px 4px rgba(156, 156, 156, 0.49);
   }
   &__img {
     height: 60%;
@@ -29,7 +37,7 @@
     left: 0;
     width: 100%;
     height: 100%;
-    object-fit: cover;
+    object-fit: contain;
     transition: all 0.2s;
     &:hover {
       scale: 1.1;
@@ -50,8 +58,8 @@
     &:before {
       content: '';
       width: 270px;
-      height: 1.5px;
-      background-color: #494949;
+      height: 1px;
+      background-color: #d9d9d9;
       position: absolute;
       top: -5px;
       left: 0;
@@ -77,12 +85,10 @@
     justify-content: center;
     align-items: center;
     font-size: 16px;
-    transition: all 0.2s;
+    transition: all 0.3s;
+    border-radius: 3px;
     &:hover {
-      transform: scale(1.17);
-      transition: all 0.2s;
-      color: black;
-      font-size: 18px;
+      opacity: 0.7;
     }
   }
