+++ src/scss/_header.scss
@@ -2,6 +2,9 @@
   height: 70px;
   display: flex;
   align-items: center;
+  box-shadow: 0px 0px 12px -2px rgba(156, 156, 156, 0.49);
+  box-shadow: 0px 0px 18px 10px rgba(0, 0, 0, 0.04);
+  background-color: white;
   &__list {
     display: flex;
     justify-content: space-between;
