+++ src/scss/_input.scss
@@ -5,7 +5,7 @@
     & input {
       position: absolute;
       height: 5px;
-      width: 100%;
+      width: 226px;
       background: none;
       pointer-events: none;
       -webkit-appearance: none;
